+++ packages/xrpl/test/utils/hashLedger.test.ts
@@ -150,22 +150,4 @@ describe('hashLedger', function () {
       'transactionHash in header does not match computed hash of transactions',
     )
   })
-
-  it('hashSignedTx - pseudo-transaction', function () {
-    const transaction: EnableAmendment = {
-      Account: 'rrrrrrrrrrrrrrrrrrrrrhoLvTp',
-      Amendment:
-        'AE35ABDEFBDE520372B31C957020B34A7A4A9DC3115A69803A44016477C84D6E',
-      Fee: '0',
-      LedgerSequence: 84206081,
-      Sequence: 0,
-      SigningPubKey: '',
-      TransactionType: 'EnableAmendment',
-    }
-
-    assert.equal(
-      hashes.hashSignedTx(transaction),
-      'CA4562711E4679FE9317DD767871E90A404C7A8B84FAFD35EC2CF0231F1F6DAF',
-    )
-  })
 })
